+++ docs/samples/browser-plugin-meetings/app.js
@@ -34,6 +34,7 @@ const integrationEnv = document.getElementById('integration-env');
 const turnDiscoveryCheckbox = document.getElementById('enable-turn-discovery');
 const eventsList = document.getElementById('events-list');
 const multistreamLayoutElm = document.querySelector('#multistream-layout');
+const preferLiveVideoElm = document.querySelector('#prefer-live-video');
 const breakoutsList = document.getElementById('breakouts-list');
 const breakoutTable = document.getElementById('breakout-table');
 const breakoutHostOperation = document.getElementById('breakout-host-operation');
@@ -1965,6 +1966,28 @@ function updateMultistreamVideoLayout() {
   }
 }
 
+function setPreferLiveVideo () {
+  let preferLiveVideo = false;
+  const meeting = getCurrentMeeting();
+
+  if (!meeting) {
+    return;
+  }
+  if (!meeting.mediaProperties.webrtcMediaConnection) {
+    return;
+  }
+  const value = preferLiveVideoElm.value;
+
+  if (value === 'Enable') {
+    preferLiveVideo = true;
+  }
+
+  if (meeting.remoteMediaManager) {
+    meeting.remoteMediaManager.setPreferLiveVideo(preferLiveVideo);
+  }
+
+}
+
 async function getStatsForVideoPane(meeting, videoPane) {
   const {remoteMedia} = videoPane;
   const {wcmeReceiveSlot} = remoteMedia.getUnderlyingReceiveSlot();
@@ -2228,6 +2251,7 @@ function addMedia() {
     // we can't import anything so can't read the initialLayoutId from the DefaultConfiguration that we're using
     // so we need to hardcode it like this:
     multistreamLayoutElm.value = 'AllEqual';
+    preferLiveVideoElm.value = 'Enable';
 
     // addMedia using the default RemoteMediaManagerConfig
     meeting.addMedia().then(() => {
@@ -2705,6 +2729,7 @@ const createBreakoutOperations = ()=>{
   const hostOperationsEl = document.createElement('div');
   let groupId = '';
   let sessionList = [];
+  let currentGroup;
   if(isHostUser && meetingsBreakoutSupportElm.checked){
     breakoutHostOperation.innerHTML = '';
     const hostOperationsTitleEl = document.createElement('h3');
@@ -2721,21 +2746,23 @@ const createBreakoutOperations = ()=>{
         breakoutTable.querySelector('table').lastChild.appendChild(tr);
       })
     }
-    const createBtn = createButton('Create Breakout Sessions', async ()=>{
+    const createGroup = async (newGroup)=>{
       await meeting.breakouts.getBreakout().then((res)=>{
-        createBtn.disabled = true;
-        deleteBtn.disabled = false;
-        startBtn.disabled = false;
+        if(!newGroup){
+          createBtn.disabled = true;
+          deleteBtn.disabled = false;
+          startBtn.disabled = false;
+        }
 
         const existedGroup = res.body.groups?.length && res.body.groups[0];
         if(existedGroup && existedGroup.status !== 'CLOSED'){
           const group = res.body.groups[0];
           const {id, sessions} = group;
           groupId = id;
-          sessionList = sessions
+          sessionList = sessions;
         }else{
           sessionList = [{'name':'session1', "anyoneCanJoin" : true}, {'name':'session2', "anyoneCanJoin" : false}];
-          meeting.breakouts.create({
+          meeting.breakouts.create(newGroup || {
             allowBackToMain: true,
             allowToJoinLater: true,
             delayCloseTime: 60,
@@ -2744,8 +2771,11 @@ const createBreakoutOperations = ()=>{
             groupId = res.body.groups[0].id;
           })
         }
-        createSessionRow();
       })
+    }
+    const createBtn = createButton('Create Breakout Sessions', async ()=>{
+      await createGroup();
+      createSessionRow();
     });
     const startBtn = createButton('Start Breakout Sessions', ()=>{
       endBtn.disabled = false;
@@ -2756,14 +2786,29 @@ const createBreakoutOperations = ()=>{
         if (!groups.length) {
           return;
         }
-        meeting.breakouts.start(groups[0]);
+        currentGroup = groups[0];
+        meeting.breakouts.start(currentGroup);
       });
     });
     const endBtn = createButton('End Breakout Sessions', ()=>{
       let countDown = meeting.breakouts.delayCloseTime;
       countDown = countDown<0?0:countDown;
       meeting.breakouts.end().then(()=>{
-        setTimeout(() => {
+        setTimeout(async () => {
+          const {sessions} = currentGroup;
+          const newSessions = sessions.map((session)=>{
+            const newSession = {...session};
+            delete newSession.id;
+            delete newSession.locusUrl;
+            return newSession;
+          })
+          const newGroup = {...currentGroup, sessions: newSessions};
+          delete newGroup.id;
+          delete newGroup.status;
+          delete newGroup.duration;
+          delete newGroup.type;
+
+          await createGroup(newGroup)
           createSessionRow();
         }, 500);
       });