+++ packages/@webex/plugin-meetings/src/locus-info/index.ts
@@ -1,4 +1,4 @@
-import {isArray, isEqual, assignWith, cloneDeep} from 'lodash';
+import {isEqual, assignWith, cloneDeep} from 'lodash';
 
 import LoggerProxy from '../common/logs/logger-proxy';
 import EventsScope from '../common/events/events-scope';
@@ -1226,18 +1226,18 @@ export default class LocusInfo extends EventsScope {
           self
         );
       }
-      // When the user upgrades to moderator or cohost
-      if (parsedSelves.updates.isUpgradeToModeratorOrCohost) {
+
+      if (parsedSelves.updates.isRolesChanged) {
         this.emitScoped(
           {
             file: 'locus-info',
             function: 'updateSelf',
           },
-          LOCUSINFO.EVENTS.SELF_MODERATOR_OR_COHOST_UPGRADE,
-          self
+          LOCUSINFO.EVENTS.SELF_ROLES_CHANGED,
+          {oldRoles: parsedSelves.previous?.roles, newRoles: parsedSelves.current?.roles}
         );
       }
-      //
+
       if (parsedSelves.updates.isVideoMutedByOthersChanged) {
         this.emitScoped(
           {